requirements.txt
qiniu==7.4.1
allure-pytest==2.8.24
allure-python-commons==2.8.24
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
idna==2.10
iniconfig==1.1.1
logger==1.4
packaging==20.7
pluggy==0.13.1
py==1.9.0
pymongo==3.11.0
PyMySQL==0.10.0
pyparsing==2.4.7
pytest==6.1.2
pytest-html==3.1.0
pytest-metadata==1.11.0
redis==3.5.3
requests==2.24.0
six==1.15.0
toml==0.10.2
urllib3==1.25.10
xlrd==1.2.0
pyyaml==5.4.1
pycryptodome==3.10.1
mongoengine
emoji==1.6.1
jsonpath==0.82
